chore: rename run args to apply args;delete unused code
