--- common/structs.go
@@ -1,31 +1,0 @@
-// Copyright © 2021 Alibaba Group Holding Ltd.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package common
-
-type RunArgs struct {
-	ClusterName string
-	Masters     string
-	Nodes       string
-	User        string
-	Password    string
-	Port        uint16
-	Pk          string
-	PkPassword  string
-	PodCidr     string
-	SvcCidr     string
-	Provider    string
-	CustomEnv   []string
-	CMDArgs     []string
-}
