--- pkg/command/cmd_proxy.go
@@ -1,53 +1,0 @@
-// Copyright © 2021 Alibaba Group Holding Ltd.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package command
-
-import (
-	"fmt"
-
-	"github.com/sealerio/sealer/common"
-)
-
-/*
-  exec some commands on remote host like create ipvs rules or add routers.
-  create certs on master1 master2.
-*/
-
-func RemoteCerts(altNames []string, hostIP, hostName, serviceCIRD, DNSDomain string) string {
-	cmd := "seautil certs "
-	if hostIP != "" {
-		cmd += fmt.Sprintf(" --node-ip %s", hostIP)
-	}
-
-	if hostName != "" {
-		cmd += fmt.Sprintf(" --node-name %s", hostName)
-	}
-
-	if serviceCIRD != "" {
-		cmd += fmt.Sprintf(" --service-cidr %s", serviceCIRD)
-	}
-
-	if DNSDomain != "" {
-		cmd += fmt.Sprintf(" --dns-domain %s", DNSDomain)
-	}
-
-	for _, name := range append(altNames, common.APIServerDomain) {
-		if name != "" {
-			cmd += fmt.Sprintf(" --alt-names %s", name)
-		}
-	}
-
-	return cmd
-}
